- Move auth.js routing file to proper directory
